+++ docutils/docutils/__init__.py
@@ -85,7 +85,7 @@ if TYPE_CHECKING:
 
 __docformat__ = 'reStructuredText'
 
-__version__ = '0.22rc3.dev'
+__version__ = '0.22rc3'
 """Docutils version identifier (complies with PEP 440)::
 
     major.minor[.micro][releaselevel[serial]][.dev]
@@ -162,7 +162,7 @@ __version_info__ = VersionInfo(
     micro=0,
     releaselevel='candidate',  # one of 'alpha', 'beta', 'candidate', 'final'
     serial=3,  # pre-release number (0 for final releases and snapshots)
-    release=False  # True for official releases and pre-releases
+    release=True  # True for official releases and pre-releases
     )
 """Comprehensive version information tuple.